requirements.txt
attrs==21.4.0
aws-cdk-lib==2.31.1
beautifulsoup4==4.11.1
cattrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
constructs==10.1.43
exceptiongroup==1.0.0rc8
idna==3.3
jsii==1.61.0
MLB-StatsAPI==1.4.2
oauthlib==3.2.0
publication==0.0.3
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
tweepy==4.10.0
typing-extensions==4.3.0
urllib3==1.26.10
